Added submission for day 10 (part 1 only)
Added day 10 part 1- which gives the correct output for the puzzle. My code for part 2 is also included- however it is not an optimal solution by any means. Although it theoretically works (tried and tested with the examples), it is so INEFFICIENT that the program runs out of heap memory due to how many paths are created on the graph along with new objects being instantiated at every node on that graph. For now, I will move onto day 11- and if time allows I will continue on with day 10 part 2, coming up with a more efficient strategy.